fix: prevent reopening menu on applicationShouldHandleReopen
