+++ MiniSim/AppDelegate.swift
@@ -20,7 +20,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     }
     
     func applicationShouldHandleReopen(_ sender: NSApplication, hasVisibleWindows flag: Bool) -> Bool {
-        miniSim.open()
         return true
     }
 }
